+++ tests/testthat/test_check_input_tbl.R
@@ -3,7 +3,7 @@ context("test check_input_tbl")
 x <- tibble::tibble(a = 1:4, b = letters[1:4], c = c(0,
   0.25, 0.5, 0.75))
 v2c <- c("a", "b", "c")
-dt2c <- c("integer", "character", "numeric")
+dt2c <- c("integer", "character", "double")
 y <- check_input_tbl(x, tbl_name = "x", parent_func = "z",
   var_to_check = v2c, dt_to_check = dt2c)
 # using default settings (all arguments NULL)
@@ -36,8 +36,7 @@ test_that("error message", {
     "ab", "abc")), "The following variables required")  # dt_to_check for dt2c is null
   expect_error(check_input_tbl(x, "x", var_to_check = v2c,
     dt_to_check = c("list", "factor", "character")))
-  # expect no error message as var_to_check set to
-  # NULL
+  # expect no error message as var_to_check set to NULL:
   expect_equal(x, check_input_tbl(x, "x", dt_to_check = c("list",
     "factor", "character")))
 })
